homework-9: fixed after review
