+++ src/store/index.js
@@ -1,10 +1,12 @@
 import {configureStore} from "@reduxjs/toolkit";
 import counterReducer from "./counter";
 import pokemonsReducer from "./pokemons";
+import boardReducer from "./board";
 
 export default configureStore({
     reducer: {
         counter: counterReducer,
         pokemons: pokemonsReducer,
+        board: boardReducer,
     }
 });
