+++ src/store/pokemons.js
@@ -7,10 +7,6 @@ export const slice = createSlice({
         isLoading: false,
         data: {},
         error: null,
-        player1Data: {},
-        player2Data: [],
-        currentPlayer: 0,
-        gameResult: null,
     },
     reducers: {
         fetchPokemons: state => ({
@@ -28,37 +24,6 @@ export const slice = createSlice({
             data: {},
             error: action.payload,
         }),
-        setPlayer1Pokemons: (state, action) => ({
-            ...state,
-            player1Data: action.payload,
-        }),
-        fetchOpponentPokemons: state => ({
-            ...state,
-            isLoading: true,
-        }),
-        fetchOpponentPokemonsResolve: (state, action) => ({
-            ...state,
-            isLoading: false,
-            player2Data: action.payload,
-        }),
-        setCurrentPlayer: (state, action) => ({
-            ...state,
-            currentPlayer: action.payload,
-        }),
-        setGameResult: (state, action) => ({
-            ...state,
-            gameResult: action.payload,
-        }),
-        clearState: state => ({
-            ...state,
-            isLoading: false,
-            data: {},
-            error: null,
-            player1Data: {},
-            player2Data: {},
-            currentPlayer: 0,
-            gameResult: null,
-        })
     }
 });
 
@@ -66,32 +31,15 @@ export const slice = createSlice({
 export const {
     fetchPokemons,
     fetchPokemonsResolve,
-    setPlayer1Pokemons,
-    fetchOpponentPokemons,
-    fetchOpponentPokemonsResolve,
-    setCurrentPlayer,
-    setGameResult,
-    clearState,
 } = slice.actions;
 
 export const pokemonsLoading = state => state.pokemons.isLoading;
 export const pokemonsData = state => state.pokemons.data;
-export const pokemonsPlayer1Data = state => state.pokemons.player1Data;
-export const pokemonsPlayer2Data = state => state.pokemons.player2Data;
-export const currentPlayer = state => state.pokemons.currentPlayer;
-export const gameResult = state => state.pokemons.gameResult;
 
 export const getPokemonsAsync = () => async (dispatch) => {
     dispatch(fetchPokemons());
     const data = await FirebaseClass.getPokemonsOnce();
     dispatch(fetchPokemonsResolve(data));
-}
-
-export const getOpponentPokemonsAsync = () => async (dispatch) => {
-    dispatch(fetchOpponentPokemons());
-    const player2Request = await fetch('https://reactmarathon-api.netlify.app/api/create-player');
-    const player2Response = await player2Request.json();
-    dispatch(fetchOpponentPokemonsResolve(player2Response.data));
 }
 
 export default slice.reducer;
